requirements.txt
aiohttp==3.5.4
bencode.py==2.0.0
bitstring==3.1.5
